Functions/Cloud Functions/Recipe_similarity/requirements.txt
google-cloud-bigquery>=2.6.0
google-cloud-firestore>=2.1.3
google-cloud-storage>=2.6.0
pandas>=1.5.2
fsspec>= 0.4.0
gcsfs>=0.4.0
pandas-gbq>=0.18.1